Splitting Range objects into their own module
